feat: custom data fetching for surveys table (#3318)
