+++ packages/lib/localStorage.ts
@@ -1,3 +1,2 @@
-export const FORMBRICKS_SURVEYS_ORIENTATION_KEY_LS = "formbricks-surveys-orientation";
 export const FORMBRICKS_SURVEYS_FILTERS_KEY_LS = "formbricks-surveys-filters";
 export const FORMBRICKS_ENVIRONMENT_ID_LS = "formbricks-environment-id";
