+++ packages/lib/survey/service.ts
@@ -277,74 +277,6 @@ export const getSurveysByActionClassId = reactCache(
     )()
 );
 
-export const getSurveysSortedByRelevance = reactCache(
-  (
-    environmentId: string,
-    limit?: number,
-    offset?: number,
-    filterCriteria?: TSurveyFilterCriteria
-  ): Promise<TSurvey[]> =>
-    cache(
-      async () => {
-        validateInputs([environmentId, ZId], [limit, ZOptionalNumber], [offset, ZOptionalNumber]);
-
-        try {
-          let surveys: TSurvey[] = [];
-          const inProgressSurveyCount = await getInProgressSurveyCount(environmentId, filterCriteria);
-
-          // Fetch surveys that are in progress first
-          const inProgressSurveys =
-            offset && offset > inProgressSurveyCount
-              ? []
-              : await prisma.survey.findMany({
-                  where: {
-                    environmentId,
-                    status: "inProgress",
-                    ...buildWhereClause(filterCriteria),
-                  },
-                  select: selectSurvey,
-                  orderBy: buildOrderByClause("updatedAt"),
-                  take: limit,
-                  skip: offset,
-                });
-
-          surveys = inProgressSurveys.map(transformPrismaSurvey);
-
-          // Determine if additional surveys are needed
-          if (offset !== undefined && limit && inProgressSurveys.length < limit) {
-            const remainingLimit = limit - inProgressSurveys.length;
-            const newOffset = Math.max(0, offset - inProgressSurveyCount);
-            const additionalSurveys = await prisma.survey.findMany({
-              where: {
-                environmentId,
-                status: { not: "inProgress" },
-                ...buildWhereClause(filterCriteria),
-              },
-              select: selectSurvey,
-              orderBy: buildOrderByClause("updatedAt"),
-              take: remainingLimit,
-              skip: newOffset,
-            });
-
-            surveys = [...surveys, ...additionalSurveys.map(transformPrismaSurvey)];
-          }
-
-          return surveys;
-        } catch (error) {
-          if (error instanceof Prisma.PrismaClientKnownRequestError) {
-            console.error(error);
-            throw new DatabaseError(error.message);
-          }
-          throw error;
-        }
-      },
-      [`getSurveysSortedByRelevance-${environmentId}-${limit}-${offset}-${JSON.stringify(filterCriteria)}`],
-      {
-        tags: [surveyCache.tag.byEnvironmentId(environmentId)],
-      }
-    )()
-);
-
 export const getSurveys = reactCache(
   (
     environmentId: string,
@@ -357,12 +289,6 @@ export const getSurveys = reactCache(
         validateInputs([environmentId, ZId], [limit, ZOptionalNumber], [offset, ZOptionalNumber]);
 
         try {
-          if (filterCriteria?.sortBy === "relevance") {
-            // Call the sortByRelevance function
-            return await getSurveysSortedByRelevance(environmentId, limit, offset ?? 0, filterCriteria);
-          }
-
-          // Fetch surveys normally with pagination
           const surveysPrisma = await prisma.survey.findMany({
             where: {
               environmentId,
